Add tested functionality to get topic and add it if non existent.
