+++ back-end/models/topicss.model.js
@@ -1,0 +1,5 @@
+const { getItemAndInsertWhenNonExistent } = require("./model-utils");
+
+exports.selectTopicAndInsertIfNonExistent = async (topic) => {
+  return await getItemAndInsertWhenNonExistent("topics", "topic_name", topic);
+};
